Separating Filters part 1
