+++ src/actions/scoreboard.js
@@ -1,6 +1,7 @@
 import { getScores } from '../APIcalls';
 
 const SET_PAST_SCORES = 'SET_PAST_SCORES'
+const SET_NUMBER_OF_CARDS = 'SET_NUMER_OF_CARDS'
 
 export const fetchPastScores = () => {
   return(dispatch) => {
@@ -12,3 +13,8 @@ export const setPastScores = (data) => ({
   type: SET_PAST_SCORES,
   payload: data
 });
+
+export const setNumberOfCardsFilter = (value) => ({
+  type: SET_NUMBER_OF_CARDS,
+  payload: value
+})
