+++ src/reducers/scoreboard.js
@@ -1,5 +1,6 @@
 const initialState = {
-  pastScores: []
+  pastScores: [],
+  numberOfCardsFilter: 'All'
 }
 
 const setPastScores = (state, data) => {
@@ -10,6 +11,8 @@ export const scoreboard = (state = initialState, action) => {
   switch(action.type) {
     case 'SET_PAST_SCORES':
       return setPastScores(state, action.payload);
+    case 'SET_NUMBER_OF_CARDS':
+      return { ...state, numberOfCardsFilter: action.payload };
     default:
       return state;
   }
